rename: ComplicateButton -> GorgeousButton 命名の指す範囲が広く中身が想像できなかったのでリネーム  #263
